requirements.txt
coreapi==2.1.1
decorator==4.0.10
Django==1.10.4
django-extensions==1.7.5
django-filter==1.0.1
django-rest-swagger==2.1.0
djangorestframework==3.5.3
ipython==5.1.0
ipython-genutils==0.1.0
itypes==1.1.0
Markdown==2.6.7
openapi-codec==1.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
requests==2.12.3
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
traitlets==4.3.1
uritemplate==3.0.0
wcwidth==0.1.7
Werkzeug==0.11.11
